b=19051 r=yong.fan, jinshan.xiong
Use 64bits (loff_t) for holding file size instead of size_t

+++ lustre/lclient/lcommon_cl.c
@@ -831,7 +831,7 @@ int ccc_prep_size(const struct lu_env *env, struct cl_object *obj,
                                  * kernel will check such case correctly.
                                  * linux-2.6.18-128.1.1 miss to do that.
                                  * --bug 17336 */
-                                size_t size = cl_isize_read(inode);
+                                loff_t size = cl_isize_read(inode);
                                 unsigned long cur_index = start >> CFS_PAGE_SHIFT;
 
                                 if ((size == 0 && cur_index != 0) ||
